有报错的地方先不要管…我暂时还写不出，明天写QvQ
更改了商品、商品分类、现金费用单。发现原来写得好水啊没有规划好QvQ

+++ src/main/java/vo/CheckCommodityItemVO.java
@@ -19,7 +19,7 @@ public class CheckCommodityItemVO {
 	/** 库存均价 */
 
 	
-	public CheckCommodityItemVO(String name, String type, int number, int aveMoney, String today, String lot) {
+	public CheckCommodityItemVO(String name, String type, int number, double aveMoney, String today, String lot) {
 		this.name = name;
 		this.type = type;
 		this.number = number;
